Move fiat-convert to subfolder
