+++ freqtrade/rpc/fiat_convert.py
